requirements.txt
deepface==0.0.75
opencv-python==4.8.0.74
Pillow==9.1.0
tk==0.1.0
cryptography==41.0.0